ajout de l actionmode et de la selection de lignes

+++ src/com/beboo/wifibackupandrestore/backupmanagement/WIFIConfigurationManager.java
@@ -127,7 +127,7 @@ public class WIFIConfigurationManager {
         String configFilename = getConfigurationFile();
         loadBackuped(backupFile.getAbsolutePath());
         loadConfigured(configFilename);
-        if (configFilename.contains("sdcard")) {
+        if (configFilename != null && configFilename.contains("sdcard")) {
             File config = new File(configFilename);
             config.delete();
         }
